+++ packages/terra-core-docs/CHANGELOG.md
@@ -11,6 +11,7 @@
   * Updated terra-button accessibility guide with anchor-button examples.
 
 * Added
+  * Added Accessibility guide for `terra-section-header`.
   * Added Hyperlink Button tests and examples.
   * Added Accessibility guide for `terra-hyperlink`.
   * Added Accessibility guide for `terra-image`.